Fix Data stall issue
In the case of disabling apn and enabling immediately, as the state
of apn type may still be CONNECTED, it will return
Phone.APN_ALREADY_ACTIVE before setting Enabled of apn type as true,
and result in the data stall.

To avoid the issue, set enabled as true in enableApnType() even it
is already connected.

Change-Id: Ie2bad4557a6bd158424625a8bb590a18d1577fc2